Display empty string for null values
